Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-event-respond-with-argument.https-expected.txt

CONSOLE MESSAGE: FetchEvent.respondWith received an error: Returned response is null.
CONSOLE MESSAGE: XMLHttpRequest cannot load https://localhost:9443/service-workers/service-worker/resources/fetch-event-respond-with-argument-iframe.html?other-value.

PASS respondWith() takes either a Response or a promise that resolves with a Response. Other values should raise a network error. 
